requirements.txt
google-generativeai
pyautogen[gemini]==0.2.25
fastapi
pyairtable
